% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{bias_correction}
\alias{bias_correction}
\title{Consensus algorithm jackknife bias correction}
\usage{
bias_correction(theta, subsamples, alpha, type = 2)
}
\arguments{
\item{theta}{a matrix containing the pooled samples from all partial 
posteriors, weighted using the consensus algorithm. See output field 
\code{theta.w.pooled} from \code{\link{consensus.weights}}.}

\item{subsamples}{a list of matrix, each consisting of (unweighted) samples 
of model parameters from a single partial posterior, chosen at random from 
all available samples from that partial posterior and representing a 
proportion \code{alpha} of all available samples. See argument \code{theta} 
to \code{\link{consensus.weights}}.}

\item{alpha}{a number in [0,1], the proportion of samples taken from all 
available from each partial posterior to form each element of 
\code{subsamples}.}

\item{type}{an integer, either 1 or 2, specifying the weighting type to use. 
1: use constant weighting. 2: weight samples using the sample covariance 
matrices of the partial posteriors. Must be the same as used to compute the 
weighted samples in \code{theta}.}
}
\value{
A list with two elements: matrix \code{theta} of bias-corrected 
pooled samples, similar to the argument of the same name, and \code{bias}, 
the bias correction vector.
}
\description{
Compute the jackknife estimate of the small sample bias that arises in the 
consensus Monte Carlo algorithm. The result can be subtracted from the 
consensus weighted sample to correct for this bias.
}
\seealso{
\code{consensus.weights}
}
