% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{consensus.weights}
\alias{consensus.weights}
\title{The consensus Monte Carlo algorithm of Scott et al 2016}
\usage{
consensus.weights(
  theta,
  type = 2,
  return.pooled = FALSE,
  correct.bias = FALSE,
  alpha = 0.2,
  par.clust = NULL,
  ncores = 1,
  cov.tol = .Machine$double.eps
)
}
\arguments{
\item{theta}{a list of matrices, each containing samples of model parameters 
from partial posterior distributions. Each matrix should have the same 
number of columns, which correspond to model parameters (including 
components of parameter vectors). They must also have the same number of 
rows (same number of samples from each partial posterior). Each list element 
corresponds to a single partial posterior.}

\item{type}{an integer, either 1 or 2, specifying the weighting type to use. 
1: use constant weighting. 2: weight samples using the sample covariance 
matrices of the partial posteriors.}

\item{return.pooled}{logical. If \code{TRUE}, pooled, weighted samples are 
returned. If Spark is used, this means returned to local memory (the calling 
environment) - in either case, weighted samples will be returned in a Spark 
table.}

\item{correct.bias}{Logical. If \code{TRUE}, the small sample bias 
correction is applied to the pooled samples. Has no effect if 
\code{return.pooled} is \code{FALSE}.}

\item{alpha}{numeric between 0 and 1: the proportion of samples to use in 
the bias estimation. Default is 0.2; ignored if \code{correct.bias} is 
\code{FALSE}.}

\item{par.clust}{an optional cluster connection object from package 
\code{parallel}. Ignored if \code{theta} is a Spark table.}

\item{ncores}{an optional integer specifying the number of CPU cores to use 
(see \code{\link[parallel]{makeCluster}}). The default, 1, signifies that 
\code{parallel} will not be used.}

\item{cov.tol}{a number, the tolerance level for the reciprocal condition 
number of the covariance matrix. If less than this, the covariances are 
ignored. See details.}
}
\value{
A list containing fields:
\item{w}{A list of weighting matrices used.}
\item{cov_used.partial}{Diagnostics (see Details).}
\item{cov_used.pooled}{Diagnostics (see Details).}
\item{theta.w.pooled}{A matrix of pooled, weighted samples if 
\code{return.pooled} is \code{TRUE}.}
\item{theta.w}{A Spark table similar to \code{theta} of weighted samples 
(if Spark is used).}
}
\description{
Compute the weights for samples from the partial posterior distributions 
according to the algorithm of Scott et al 2016, resulting in a pooled, 
weighted sample that can be used to approximate expected values under the 
full data posterior distribution.
}
\details{
When \code{theta} is a list, \code{consensus.weights} replicates on a single 
machine what would be performed on a cluster. Each element of \code{theta} 
would correspond to the samples from a single partial posterior distribution.

This is implemented in a distributed manner using Spark when \code{theta} is 
a Spark table.

Parameter samples \code{theta} should be sampled from the partial posterior 
distributions using "fractionated" priors as in Scott et al 2016. This is 
assumed to be the case.

Output fields \code{cov_used.partial} and \code{cov_used.pooled} indicate 
whether we used the full covariance matrices of the partial or pooled 
samples, respectively, for weighting. In some cases the inverse covariance 
matrix is difficult to compute, in which case the covariances are ignored, 
as suggested by Scott et al 2016.
}
\section{References}{

\itemize{
\item{Scott, Steven L., Blocker, A.W., Bonassi, F.V., Chipman, H.A., George, E.I. and McCulloch, R.E., 2016. Bayes and big data: The consensus Monte Carlo algorithm. \emph{International Journal of Management Science and Engineering Management}, 11(2), pp.78-88.}
}
}

