% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.r
\name{consensus.worker}
\alias{consensus.worker}
\title{Compute consensus algorithm weight matrices}
\usage{
consensus.worker(df, context)
}
\arguments{
\item{df}{a matrix or data.frame of samples. If using Spark the first column 
should contain integers identifying the partial posterior.}

\item{context}{a list of parameters common to all workers. See details.}
}
\value{
Consensus algorithm weighting matrix for the sample. If using Spark 
the first column will contain the index of the partial posterior from which 
the sample was drawn. The next column is a binary flag indicating whether 
(1) or not (0) the full covariance matrix was used. The number of remaining 
columns is equal to the number of rows. These columns constitute a square 
matrix: the weighting matrix for consensus algorithm \code{type}.
}
\description{
This is a worker function for computing weight matrices for samples from 
partial posteriors. If samples are held in local memory, should be applied 
iteratively using a loop, but can be applied in parallel using the 
\code{parallel} package or \code{sparklyr} if samples are distributed as a 
Spark table.
}
\details{
\code{context} should be a list with the following fields:
\describe{
\item{\code{type}}{Integer, either 1 or 2, specifying the weighting type to 
use. 1: use constant weighting. 2: weight samples using the sample covariance 
matrices of the partial posteriors.}
\item{\code{use_spark}}{Logical. \code{TRUE} if a Spark cluster is available.}
\item{\code{tol}}{The tolerance level for the reciprocal condition number of 
the covariance matrix. If less than this, the covariances are ignored.}
}
}
\section{References}{

\itemize{
\item{Scott, Steven L., Blocker, A.W., Bonassi, F.V., Chipman, H.A., George, E.I. and McCulloch, R.E., 2016. Bayes and big data: The consensus Monte Carlo algorithm. \emph{International Journal of Management Science and Engineering Management}, 11(2), pp.78-88.}
}
}

