% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemie.r
\name{lemie.mean}
\alias{lemie.mean}
\title{Monte Carlo estimate of the expectation of a univariate function}
\usage{
lemie.mean(theta, wn, FUN = identity, type = 2, Hvec = NULL)
}
\arguments{
\item{theta}{a list of matrices, each containing samples of model parameters 
from partial posterior distributions. Each matrix should have the same 
number of columns, which correspond to model parameters (including 
components of parameter vectors). Each list element corresponds to a single 
partial posterior.}

\item{wn}{a list of matrices (for \code{type = 1} or \code{type = 2}) or a 
matrix (\code{type = 3}), each containing normalised weights, one for each 
sample in \code{theta} and obtained using the \code{\link{lemie.weights}} 
function.}

\item{FUN}{a matrix-valued function that we wish to estimate the expected 
value of. See details.}

\item{type}{an integer, either 1, 2 or 3, specifying the weighting type used.}

\item{Hvec}{an optional vector specifying the number of samples taken from 
each partial posterior.}
}
\value{
A vector of weighted sample means of \code{FUN}, approximating the 
posterior expectation.
}
\description{
Compute a Monte Carlo estimate of the expectation of a function of model 
parameters. Samples of the parameter vector are weighted using the LEMIE 
type 1, 2 or 3 importance weights.
}
\details{
\code{FUN} should take a matrix argument and return a matrix. The argument 
matrix should be a matrix of samples, just like the elements of 
\code{theta}. The returned matrix can have any number of columns but should 
have the same number of rows. The rows of both matrices correspond to 
samples.
}
\seealso{
\code{lemie.quantile}, \code{lemie.weights}
}
