% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemie.r
\name{lemie.normalise}
\alias{lemie.normalise}
\title{Normalise LEMIE weights}
\usage{
lemie.normalise(
  w,
  type,
  pp.inds = NULL,
  just_compute_constant = FALSE,
  as.list = TRUE,
  par.clust = NULL,
  forking = FALSE,
  ncores = 1,
  verbose = FALSE
)
}
\arguments{
\item{w}{either a single column matrix, if \code{type = 2} or 
\code{type = 3}, or a list of single column matrices, if \code{type = 1}, 
containing the unnormalised LEMIE weights on the log scale.}

\item{type}{an integer, either 1, 2 or 3, specifying the weighting type used.}

\item{pp.inds}{integer vector, required for type 2 if \code{as.list} is 
\code{TRUE} (default), with one element for each row of \code{w}: the index 
of the partial posterior the corresponding sample is from.}

\item{just_compute_constant}{Logical. If \code{TRUE}, only the normalising 
constant(s) will be computed and returned - not the normalised weights.}

\item{as.list}{logical. If \code{TRUE} the normalised type 2 weights will be 
returned as a list, split according to \code{pp.inds}.}

\item{par.clust}{an optional cluster connection object from package 
\code{parallel}.}

\item{forking}{logical. If \code{TRUE}, use forking functions 
\code{\link[parallel]{mclapply}}, \code{\link[parallel]{mcmapply}}. Does not 
work on Windows.}

\item{ncores}{an optional integer specifying the number of CPU cores to use 
(see \code{\link[parallel]{makeCluster}}). The default, 1, signifies that 
\code{parallel} will not be used.}
}
\value{
A list containing fields:
\item{wn}{A list of single column matrices, if \code{type = 1}, containing 
the normalised LEMIE weights on the log scale. If \code{as.list} is 
\code{FALSE} and \code{type} is 2, these will be returned as a single 
matrix. If \code{type}, a single column matrix.}
\item{w.sum}{Either a numeric, if \code{type = 2} or \code{type = 3}, or a 
list of numerics, if \code{type = 1}, reporting the normalising constants 
used (log scale).}
}
\description{
Performs the self-normalisation of the importance weights in LEMIE, 
necessary for estimating posterior expectations with the weighted samples 
from partial posteriors.
}
\details{
Weight normalisation is done differently for the two versions of LEMIE, as 
specified by the \code{type} argument.

The type 1 weights will be normalised relative to the partial posterior from 
which the corresponding samples were drawn. This means if there are $M$ 
partial posteriors, the sum of the normalised weights will be $M$.
}
