% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemie.r
\name{lemie.quantile}
\alias{lemie.quantile}
\title{Monte Carlo estimate of a quantile of a univariate function}
\usage{
lemie.quantile(
  theta,
  prob,
  wn,
  FUN = identity,
  type = 2,
  tol = .Machine$double.eps^0.5,
  Hvec = NULL
)
}
\arguments{
\item{theta}{a list of matrices, each containing samples of model parameters 
from partial posterior distributions. Each matrix should have the same 
number of columns, which correspond to model parameters (including 
components of parameter vectors). Each list element corresponds to a single 
partial posterior.}

\item{prob}{a probability, in [0, 1], corresponding to the quantile of 
\code{FUN} to be estimated.}

\item{wn}{a list of matrices (for \code{type = 1} or \code{type = 2}) or a 
matrix (\code{type = 3}), each containing normalised weights, one for each 
sample in \code{theta} and obtained using the \code{\link{lemie.weights}} 
function.}

\item{FUN}{a matrix-valued function that we wish to estimate the expected 
value of. See details.}

\item{type}{an integer, either 1, 2 or 3, specifying the weighting type used.}

\item{tol}{a number specifying the tolerance level for convergence of 
numerical root finding. See \code{\link{uniroot}} for details.}

\item{Hvec}{an optional vector specifying the number of samples taken from 
each partial posterior.}

\item{A}{matrix with named rows matching the output of 
\code{\link{uniroot}}, the first row of which contains the weighted 
quantiles of \code{FUN}.}
}
\description{
Compute a Monte Carlo estimate of quantiles of a function of model 
parameters. Samples of the parameter vector are weighted using the LEMIE 
type 1 or type 2 importance weights.
}
\details{
Whilst \code{FUN} is matrix-valued, with possibly >1 columns, the estimated 
quantiles returned are quantiles of each dimension of the function 
separately, i.e. the marginals.

\code{FUN} should take a matrix argument and return a matrix. The argument 
matrix should be a matrix of samples, just like the elements of 
\code{theta}. The returned matrix can have any number of columns but should 
have the same number of rows. The rows of both matrices correspond to 
samples.
}
\seealso{
\code{lemie.mean}, \code{lemie.weights}
}
