% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{logistic.sampler}
\alias{logistic.sampler}
\title{Gibbs Sampling in a Logistic Regression Model}
\usage{
logistic.sampler(
  y,
  x,
  weights = rep(1, nrow(x)),
  offset = NULL,
  H = 1000,
  params = list()
)
}
\arguments{
\item{y}{integer or logical vector, the dependent variable.}

\item{x}{matrix of independent variables. Each column is interpreted as a 
predictor variable.}

\item{weights}{an optional vector of sample weights to use in a binomial 
logistic regression where the dependent variable \code{y} is an integer 
vector. In this case \code{weights} are the corresponding case weights.}

\item{offset}{optional vector of offsets.}

\item{H}{an integer specifying the number of MCMC iterations and therefore 
of samples to draw.}

\item{params}{list of model parameters. See details.}
}
\value{
A list containing matrix of samples, \code{samples}, and list 
\code{params}.
}
\description{
Samples from the posterior distribution of model parameters (coefficients) 
in a logistic regression model with the logit link and using supplied data. 
This Gibbs sampling algorithm is due to Polson et al 2013.
}
\details{
The model is specified via argument \code{params} (also returned as a field 
of the output list). \code{params} is a list with the following fields:

\describe{
\item{\code{mu}}{Location parameter vector in multivariate normal prior for 
coefficients.}
\item{\code{Sigma}}{Covariance matrix in multivariate normal prior for 
coefficients.}
}
}
\section{References}{

\itemize{
\item{Polson, N.G., Scott, J.G. and Windle, J., 2013. Bayesian inference for logistic model using P\out{&oacture;}lya-Gamma latent variables. \emph{Journal of the American statistical Association, 108} \bold{(504)}, pp.1339-1349.}
}
}

