% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parallel.start}
\alias{parallel.start}
\title{Verify an existing or initialise a new parallel cluster}
\usage{
parallel.start(par.clust = NULL, ncores = 1)
}
\arguments{
\item{par.clust}{an optional cluster connection object from package 
\code{parallel}.}

\item{ncores}{an optional integer specifying the number of CPU cores to use 
(see \code{\link[parallel]{makeCluster}}).}
}
\value{
A list containing fields:
\item{par.clust}{Cluster connection object from package \code{parallel}.}
\item{valid}{Logical. \code{TRUE} if \code{par.clust} is working.}
\item{new}{Logical. \code{TRUE} if \code{par.clust} was created by this 
function.}
\item{ncores}{Integer. The number of cores used by \code{par.clust}.}
}
\description{
This is a wrapper of \code{parallel::detectCores} and 
\code{parallel::makeCluster}.
}
