% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemie.r
\name{pareto_smooth}
\alias{pareto_smooth}
\title{Smooth the LEMIE weights using the generalised Pareto distribution}
\usage{
pareto_smooth()
}
\arguments{
\item{psis}{logical. If \code{TRUE} the Pareto smoothed importance sampling 
(PSIS) algorithm of Vehtari et al (2015) is applied to the raw importance 
weights.}
}
\description{
Smooth the LEMIE weights using the generalised Pareto distribution
}
\section{References}{

\itemize{
\item{Vehtari, A., Simpson, D., Gelman, A. Yao, Y. and Gabry, J., 2015. Pareto smoothed importance sampling. \emph{arXiv preprint arXiv: 1507.02646.}}
\item{Vehtari, A., Gelman, A. and Gabry, J., 2017. Practical Bayesian model evaluation using leave-one-out cross-validation and WAIC. \emph{Statistics and computing}, 27(5), pp.1413-1432.}
}
}

