% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partition}
\alias{partition}
\title{Partition matrix-like data into \code{r} parts}
\usage{
partition(
  x,
  r = 2,
  part = NULL,
  part_weights = rep(1, r),
  random = TRUE,
  balanced = FALSE
)
}
\arguments{
\item{x}{a vector, matrix or data.frame of data to be partitioned.}

\item{r}{an integer, the required number of parts.}

\item{part}{a vector the same length as \code{x} of part labels that can be 
used to specify how to partition \code{x}.}

\item{part_weights}{a vector of length \code{r} specifying the relative 
sizes of the parts or probability of part membership in the unbalanced 
random allocation.}

\item{random}{logical. If \code{TRUE}, elements of \code{x} are assigned to 
parts at random.}

\item{balanced}{logical. If \code{TRUE}, and if \code{random} is 
\code{TRUE}, the parts will have the same number of elements (or as close as 
possible). \code{part_weights} will be ignored.}
}
\value{
A list of nonintersecting subsets of \code{x}.
}
\description{
Performs one of the following partitions: sequential (the first \code{n_i} 
elements go into part \code{i}, etc), unbalanced random (random allocation 
into parts which are allowed to have different size) or balanced random 
(random allocation into parts of the same size).
}
\details{
If \code{x} is a matrix the partition is of sets of rows.

Alternatively an integer scalar \code{x} can be supplied, in which case the 
partition is of the integers \code{1:x}.
}
