% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{scott_logistic}
\alias{scott_logistic}
\title{Artificial Data Set for Logistic Regression from Scott et al 2016}
\format{
A data.frame with 10,000 rows and 6 variables:
\describe{
\item{y}{Numeric; outcome}
\item{x1}{Numeric}
\item{x2}{Numeric}
\item{x3}{Numeric}
\item{x4}{Numeric}
\item{x5}{Numeric}
}
}
\source{
Scott, Steven L., Blocker, A.W., Bonassi, F.V., Chipman, H.A., George, E.I. and McCulloch, R.E., 2016. Bayes and big data: The consensus Monte Carlo algorithm. \emph{International Journal of Management Science and Engineering Management}, 11(2), pp.78-88.
}
\usage{
scott_logistic
}
\description{
A data set used to test the consensus Monte Carlo algorithm in Scott et al 
2016,  containing 10,000 records of binary responses and 5 binary predictors.
}
\keyword{datasets}
