% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{parallel.start}
\alias{parallel.start}
\title{Verify an existing or initialise a new parallel cluster}
\usage{
parallel.start(par.clust = NULL, ncores = 1, forking = FALSE)
}
\arguments{
\item{par.clust}{an optional cluster connection object from package 
\code{parallel}.}

\item{ncores}{an optional integer specifying the number of CPU cores to use 
(see \code{\link[parallel]{makeCluster}}).}

\item{forking}{logical. If \code{TRUE}, forking functions 
\code{\link[parallel]{mclapply}}, \code{\link[parallel]{mcmapply}} will be 
used and no connection object will be defined.}
}
\value{
A list containing fields:
\item{par.clust}{Cluster connection object from package \code{parallel}.}
\item{valid}{Logical. \code{TRUE} if \code{par.clust} is working.}
\item{new}{Logical. \code{TRUE} if \code{par.clust} was created by this 
function.}
\item{ncores}{Integer. The number of cores used by \code{par.clust}.}
}
\description{
This is a wrapper of \code{parallel::detectCores} and 
\code{parallel::makeCluster}.
}
